requirements.txt
sqlalchemy>=2.0
pymysql>=1.1
requests>=2.31
beautifulsoup4>=4.12
selenium>=4.18
webdriver-manager>=4.0
pandas>=2.2
thefuzz>=0.22
python-dotenv>=1.0
statsbombpy>=1.1
lxml>=5.1
soccerdata>=1.4,<2
